Updated glutin and switched to the gfx git master

+++ src/context/mod.rs
@@ -47,9 +47,9 @@ impl Context {
     }
 
     // Resize the context
-    pub fn resize(&mut self, width: f32, height: f32) {
-        self.width = width;
-        self.height = height;
+    pub fn resize(&mut self, width: u32, height: u32) {
+        self.width = width as f32;
+        self.height = height as f32;
         // resize the renderer
         self.renderer.resize(width, height);
     }
